+++ tests/js/session-tests.js
@@ -105,14 +105,11 @@ module.exports = {
             return;
         }
 
-        const username = uuid();
-        const realmName = uuid();
-
-        return Realm.Sync.User.register('http://localhost:9080', username, 'password').then(user => {
+        return Realm.Sync.User.login('http://localhost:9080', Realm.Sync.Credentials.anonymous()).then(user => {
             let config = {
-                sync: { 
-                    user, 
-                    url: `realm://localhost:9080/~/${realmName}`, 
+                sync: {
+                    user,
+                    url: `realm://localhost:9080/~/${uuid()}`,
                     fullSynchronization: true,
                     custom_http_headers: {
                         'X-Foo': 'Bar'
@@ -131,7 +128,7 @@ module.exports = {
     },
 
     testProperties() {
-        return Realm.Sync.User.register('http://localhost:9080', uuid(), 'password').then(user => {
+        return Realm.Sync.User.login('http://localhost:9080', Realm.Sync.Credentials.anonymous()).then(user => {
             return new Promise((resolve, reject) => {
                 const accessTokenRefreshed = this;
                 let successCounter = 0;
@@ -180,12 +177,12 @@ module.exports = {
         const expectedObjectsCount = 3;
 
         let user, config;
+
+        const credentials = Realm.Sync.Credentials.nickname(username);
         return runOutOfProcess(__dirname + '/download-api-helper.js', username, realmName, REALM_MODULE_PATH)
-            .then(() => Realm.Sync.User.login('http://localhost:9080', username, 'password'))
+            .then(() => Realm.Sync.User.login('http://localhost:9080', credentials))
             .then(u => {
                 user = u;
-                const accessTokenRefreshed = this;
-                let successCounter = 0;
 
                 config = {
                     sync: { user, url: `realm://localhost:9080/~/${realmName}`, fullSynchronization: true },
@@ -216,13 +213,11 @@ module.exports = {
         const expectedObjectsCount = 3;
 
         let user, config;
+        const credentials = Realm.Sync.Credentials.nickname(username);
         return runOutOfProcess(__dirname + '/download-api-helper.js', username, realmName, REALM_MODULE_PATH)
-            .then(() => Realm.Sync.User.login('http://localhost:9080', username, 'password'))
+            .then(() => Realm.Sync.User.login('http://localhost:9080', credentials))
             .then(u => {
                 user = u;
-                const accessTokenRefreshed = this;
-                let successCounter = 0;
-
                 config = {
                     sync: { user, url: `realm://localhost:9080/~/${realmName}`, fullSynchronization: true },
                     schema: [{ name: 'Dog', properties: { name: 'string' } }],
@@ -259,12 +254,10 @@ module.exports = {
         const realmName = uuid();
         const expectedObjectsCount = 3;
 
+        const credentials = Realm.Sync.Credentials.nickname(username);
         return runOutOfProcess(__dirname + '/download-api-helper.js', username, realmName, REALM_MODULE_PATH)
-            .then(() => Realm.Sync.User.login('http://localhost:9080', username, 'password'))
+            .then(() => Realm.Sync.User.login('http://localhost:9080', credentials))
             .then(user => {
-                const accessTokenRefreshed = this;
-                let successCounter = 0;
-
                 let config = {
                     sync: { user, url: `realm://localhost:9080/~/${realmName}`, fullSynchronization: true },
                     schema: [{ name: 'Dog', properties: { name: 'string' } }],
@@ -310,12 +303,10 @@ module.exports = {
         const realmName = uuid();
         const expectedObjectsCount = 3;
 
-       return runOutOfProcess(__dirname + '/download-api-helper.js', username, realmName, REALM_MODULE_PATH)
-            .then(() => Realm.Sync.User.login('http://localhost:9080', username, 'password'))
+        const credentials = Realm.Sync.Credentials.nickname(username);
+        return runOutOfProcess(__dirname + '/download-api-helper.js', username, realmName, REALM_MODULE_PATH)
+            .then(() => Realm.Sync.User.login('http://localhost:9080', credentials))
             .then(user => {
-                const accessTokenRefreshed = this;
-                let successCounter = 0;
-
                 let config = {
                     sync: { user, url: `realm://localhost:9080/~/${realmName}`, fullSynchronization: true }
                 };
@@ -410,7 +401,7 @@ module.exports = {
     },
 
     testErrorHandling() {
-        return Realm.Sync.User.register('http://localhost:9080', uuid(), 'password').then(user => {
+        return Realm.Sync.User.login('http://localhost:9080', Realm.Sync.Credentials.anonymous()).then(user => {
             return new Promise((resolve, _reject) => {
                 const config = user.createConfiguration({ sync: { url: 'realm://localhost:9080/~/myrealm' } });
                 config.sync.error = (sender, error) => {
@@ -440,8 +431,9 @@ module.exports = {
         const username = uuid();
         const realmName = uuid();
 
+        const credentials = Realm.Sync.Credentials.nickname(username);
         return runOutOfProcess(__dirname + '/nested-list-helper.js', __dirname + '/schemas.js', username, realmName, REALM_MODULE_PATH)
-            .then(() => Realm.Sync.User.login('http://localhost:9080', username, 'password'))
+            .then(() => Realm.Sync.User.login('http://localhost:9080', credentials))
             .then(user => {
                 let config = {
                     schema: [schemas.ParentObject, schemas.NameObject],
@@ -481,7 +473,7 @@ module.exports = {
             Realm.copyBundledRealmFiles();
         }
 
-        return Realm.Sync.User.register('http://localhost:9080', uuid(), 'password')
+        return Realm.Sync.User.login('http://localhost:9080', Realm.Sync.Credentials.anonymous())
             .then(user => {
                 const config = {
                     path: realm,
@@ -524,7 +516,7 @@ module.exports = {
             Realm.copyBundledRealmFiles();
         }
 
-        return Realm.Sync.User.register('http://localhost:9080', uuid(), 'password').then(user => {
+        return Realm.Sync.User.login('http://localhost:9080', Realm.Sync.Credentials.anonymous()).then(user => {
             return new Promise((resolve, _reject) => {
                 const config = {
                     path: realm,
@@ -565,7 +557,7 @@ module.exports = {
             Realm.copyBundledRealmFiles();
         }
 
-        return Realm.Sync.User.register('http://localhost:9080', uuid(), 'password').then(user => {
+        return Realm.Sync.User.login('http://localhost:9080', Realm.Sync.Credentials.anonymous()).then(user => {
             return new Promise((resolve, _reject) => {
                     const config = {
                         path: realm,
@@ -636,8 +628,9 @@ module.exports = {
         const username = uuid();
         const realmName = uuid();
 
+        const credentials = Realm.Sync.Credentials.nickname(username);
         return runOutOfProcess(__dirname + '/download-api-helper.js', username, realmName, REALM_MODULE_PATH)
-            .then(() => Realm.Sync.User.login('http://localhost:9080', username, 'password'))
+            .then(() => Realm.Sync.User.login('http://localhost:9080', credentials))
             .then(user => {
                 let config = {
                     sync: {
@@ -703,8 +696,9 @@ module.exports = {
         const realmName = uuid();
         let progressCalled = false;
 
+        const credentials = Realm.Sync.Credentials.nickname(username);
         return runOutOfProcess(__dirname + '/download-api-helper.js', username, realmName, REALM_MODULE_PATH)
-            .then(() => Realm.Sync.User.login('http://localhost:9080', username, 'password'))
+            .then(() => Realm.Sync.User.login('http://localhost:9080', credentials))
             .then(user => {
                 let config = {
                     sync: {
@@ -729,8 +723,9 @@ module.exports = {
         const username = uuid();
         const realmName = uuid();
 
+        const credentials = Realm.Sync.Credentials.nickname(username);
         return runOutOfProcess(__dirname + '/download-api-helper.js', username, realmName, REALM_MODULE_PATH)
-            .then(() => Realm.Sync.User.login('http://localhost:9080', username, 'password'))
+            .then(() => Realm.Sync.User.login('http://localhost:9080', credentials))
             .then(user => {
                 return new Promise((resolve, reject) => {
                     let config = {
@@ -836,9 +831,10 @@ module.exports = {
             TestCase.assertThrows(() => { let realm = new Realm(config); } );
         }
 
+        const credentials = Realm.Sync.Credentials.nickname(username);
         return runOutOfProcess(__dirname + '/partial-sync-api-helper.js', username, REALM_MODULE_PATH)
             .then(() => {
-                return Realm.Sync.User.login('http://localhost:9080', username, 'password').then((u) => {
+                return Realm.Sync.User.login('http://localhost:9080', credentials).then((u) => {
                     user = u;
 
                     __partialIsAllowed();
@@ -921,7 +917,7 @@ module.exports = {
             return;
         }
 
-        return Realm.Sync.User.register('http://localhost:9080', uuid(), 'password').then(user => {
+        return Realm.Sync.User.login('http://localhost:9080', Realm.Sync.Credentials.anonymous()).then(user => {
             return new Promise((resolve, _reject) => {
                 var realm;
                 const config = user.createConfiguration({ sync: { url: 'realm://localhost:9080/~/myrealm' } });
@@ -954,7 +950,7 @@ module.exports = {
             return;
         }
 
-        return Realm.Sync.User.register('http://localhost:9080', uuid(), 'password').then((u) => {
+        return Realm.Sync.User.login('http://localhost:9080', Realm.Sync.Credentials.anonymous()).then((u) => {
             return new Promise((resolve, reject) => {
                 let config = {
                     sync: {
@@ -963,7 +959,7 @@ module.exports = {
                         fullSynchronization: true,
                     }
                 };
-                
+
                 Realm.open(config).then(realm => {
                     realm.syncSession.addConnectionNotification((newState, oldState) => {
                         if (oldState === Realm.Sync.ConnectionState.Connected && newState === Realm.Sync.ConnectionState.Disconnected) {
@@ -981,7 +977,7 @@ module.exports = {
             return;
         }
 
-        return Realm.Sync.User.register('http://localhost:9080', uuid(), 'password').then((u) => {
+        return Realm.Sync.User.login('http://localhost:9080', Realm.Sync.Credentials.anonymous()).then((u) => {
             return new Promise((resolve, reject) => {
                 let config = {
                     sync: {
@@ -990,7 +986,7 @@ module.exports = {
                         fullSynchronization: true,
                     }
                 };
-                
+
                 Realm.open(config).then(realm => {
                     let callback1 = () => {
                         reject("Should not be called");
@@ -1015,7 +1011,7 @@ module.exports = {
             return;
         }
 
-        return Realm.Sync.User.register('http://localhost:9080', uuid(), 'password').then((u) => {
+        return Realm.Sync.User.login('http://localhost:9080', Realm.Sync.Credentials.anonymous()).then((u) => {
             return new Promise((resolve, reject) => {
                 let config = {
                     sync: {
@@ -1171,7 +1167,7 @@ module.exports = {
             return;
         }
 
-        return Realm.Sync.User.register('http://localhost:9080', uuid(), 'password').then((u) => {
+        return Realm.Sync.User.login('http://localhost:9080', Realm.Sync.Credentials.anonymous()).then((u) => {
             return new Promise((resolve, reject) => {
                 let realm = new Realm(u.createConfiguration());
                 TestCase.assertEqual(5, realm.objects(Realm.Permissions.Class.schema.name).length);